tests: Add a missing test for post-normalisation lowercasing in IDN
See https://gitlab.gnome.org/GNOME/glib/-/merge_requests/4387#note_2269324

This adds a test to increase the code coverage of `nameprep()` in
`ghostutils.c`. It was previously missing coverage of the second
`tolower()` operation. The new test triggers this by using a Unicode
codepoint which cannot be converted to lowercase itself, but which
normalises (NFKC) to uppercase characters which can — a Unicode Roman
numeral.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>