gdate: Add week-of-year APIs for weeks starting on any day
`GDate` currently has some API which is specific to which day of the week
you think the week starts on:
 * `g_date_get_monday_week_of_year()`
 * `g_date_get_monday_weeks_in_year()`
 * `g_date_get_sunday_week_of_year()`
 * `g_date_get_sunday_weeks_in_year()`

This is to deal with the difference between locales which think that the
week starts on a Sunday (such as `LANG=en_US.utf8` or `LANG=he_IL.utf8`),
or a Monday (such as `LANG=en_GB.utf8`).

However, there are some locales which think that the week starts on a
Saturday (such as `LANG=ar_EG.utf8`). Currently, GLib provides no API for
those.

So, add some API which is parameterised by the first day of the week,
which will deal with weeks which start on a Saturday (and also any other
day of the week, although I don’t believe there are any countries which
use a day other than Saturday, Sunday or Monday).

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #3617