gdataset: avoid initializing temporary array with NULL in g_data_remove_internal()
The code comment was correct in the past, and how the code was then we
had to ensure the array is initialized.

This is no longer the case. We will only access the array at the indexes
from zero to up to (found_keys-1). And those places get always
initialized before we use them (as we increment found_keys at the same
time).

There is no longer a need to initialize this array to zero. Drop this.

Fixes: 3437414dd4f7 ('gdataset: use lookup index in g_datalist_id_remove_multiple()')