gdataset: avoid initializing "out_idx" if datalist_find() fails
In some sense, it might be desirable to always initialize out parameters
to something.

However, this fallback index G_MAXUINT32 is not really useful for any
caller. That is, no caller can blindly access the index without checking
whether datalist_find() found (and returned) an entry.

Also, a caller cannot check for whether a entry was found via `if (idx
== G_MAXUINT32)`, because that theoretically is also a valid index.

The only sensible way to use datalist_find() is by accessing "idx"
if-and-only-if an entry was found. And all callers did that.

There is no point for initializing the value.

Uninitialized values can be even preferable over a non-useful
initialization to G_MAXUINT32, because if there is a buggy access them
under valgrind, there will be at least a warning.