gobject/performance: only print test message in verbose mode
Historically, there was a verbose mode and a non-verbose mode.
In non-verbose mode (the default), we would still print two lines:

  Running test property-set
  Property set per second: 39329344

Later, this was changed to include the test name in the second line, so
we would print:

  Running test property-set
  property-set: Property set per second: 39329344

But this first line is really just noise, making parsing and reading the
results harder. Hence a "--quiet" mode was added, that only printed one
line per test while keeping the previous default behavior. And all was
good.

Except, unless you want verbose mode, this "Running test" line is still
not very useful and mainly clutters the output.

Supporess it now also in normal mode. It is now only printed in verbose
mode.

This also makes the "--quiet" option do nothing. The option is still
there, maybe we find a future use and we should not break the command
line API by dropping an argument.