gsettings: Clarify settings types in docs
Since the docs are saying what type a key must be in the schema to be
able to call that method, it makes sense to give the type in the same
format used in the schema, i.e. a GVariant type string.

Also link to the `GVariantType` documentation so the user can read up on
it further if needed.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>