gobject: preserve weak notifications registered during dispose
We call g_object_weak_release_all() at two places.

Once right before finalize(). At this point, the object is definitely
going to be destroyed, and the user must no longer resurrect it or
subscribe new weak notifications. In that case, we really want to
notify/release all weak notifications.

However, we also call it from g_object_real_dispose(). During dispose,
the API allows the user to resurrect an object. Granted, that is
probably not something anybody should do, but GObject makes a reasonable
attempt to support that.

If the GObject implementation itself is the one that resurrects the
object, it could just not chain up the parent implementation of
dispose(), take a strong reference and be done.

How would an external user resurrect an object? The only place is when
GObject calls g_object_real_dispose().

  static void
  g_object_real_dispose (GObject *object)
  {
    g_signal_handlers_destroy (object);

    /* GWeakNotify and GClosure can call into user code */
    g_object_weak_release_all (object, FALSE);
    closure_array_destroy_all (object);
  }

Closures aside, the place to resurrect the object is is the weak
notification.

But previously, g_object_weak_release_all() would continue iterating
until there are no more weak notifications left. So while the user can
take a strong reference and resurrect the object, their attempts to
register new weak notifications are thwarted.

Instead, when the loop in g_object_weak_release_all() starts, remember
the initial number of weak notifications, and don't release more than
that. Note that WeakRefStack preserves the order of entries now, so by
maintaining the "remaining_to_notify" counter we know when to stop.

Note that this brings also the previous behavior back, where we would
call

  g_datalist_id_set_data (&object->qdata, quark_weak_notifies, NULL);

This would take out the WeakRefStack and notify the ones registered at
the time, but subsequent registrations would not be notified yet.