gobject/docs: comment what g_object_run_dispose() does
It seems relevant how g_object_run_dispose() handles #GWeakRef.
It's not clear when the best time is to clear them.

For example, if the dispose() method wishes to register new #GWeakRef,
then those would be cleared right away too, even if the object does
not actually get destroyed. Should we thus clear #GWeakRef before
dispose()? Or not at all?

Unclear. But it is important for the user to understand how this works.

Document it thus.