gobject: clear weak locations before calling dispose in g_object_run_dispose()
This changes behavior from commit [1].

The point of g_object_run_dispose() is to break reference cycles to
bring down an object. We don't expect the object to take new references
to keep it alive for longer. We probably also don't expect it to
register new weak references. We also don't expect the dispose()
callees to check g_weak_ref_get() for the object. In those situations,
this change makes no difference.

I wonder, whether we should clear GWeakRef at all, because during
g_object_run_dispose(), the real reference counter is not yet
going to NULL. Maybe the GWeakRef should honor that primarily, and
not clear out at g_object_run_dispose(). On the other hand, dispose()
will also emit weak notifications too, despite the object is not yet
going away for real. So maybe we indeed want this. And this is what we
did since [1] and before.

But compare to g_object_unref(), where we also clear GWeakRef *before*
calling dispose. That makes sense, because inside dispose() (and for
example during weak notifications), we probably want to see that
g_weak_ref_get() indicates the object is gone. For that reason, it seems
more correct to clear out the GWeakRef before calling dispose().

Also, the dispose() callees (e.g. the weak notifications) might refuse to
let the object die by intentionally keeping strong references around.
Not sure why they would do that, it is similar to resurrecting an object
during dispose(). But if they do, they might also want to register new
GWeakRef. In that case, we wouldn't want to unset those newly set
GWeakRef unconditionally right after.

In most cases, it shouldn't make a difference. In the case where it
does, this seems the more sensible order of things.

[1] commit 2952cfd7a7f2 ('gobject: drop clearing quark_weak_locations from g_object_real_dispose()')